+++ NEWS.md
@@ -1,3 +1,5 @@
+# rsimsum 0.6.0-9000
+
 # rsimsum 0.6.0
 
 ### Improvements: